Move connectivity tool out of orders package
The connectivity tool will be accessible from Settings in addition
to the order list, so it no longer belongs under the orders package.

- Move ui/orders/connectivitytool/ → ui/connectivitytool/
- Rename OrderConnectivityToolFragment → ConnectivityToolFragment
- Rename OrderConnectivityToolViewModel → ConnectivityToolViewModel
- Rename OrderConnectivityToolScreen → ConnectivityToolScreen
- Update nav graph and OrderListFragment references
