+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/list/OrderListFragment.kt
@@ -974,7 +974,7 @@ class OrderListFragment :
 
     private fun openConnectivityTool() {
         findNavController().navigateSafely(
-            OrderListFragmentDirections.actionOrderListFragmentToOrderConnectivityToolFragment()
+            OrderListFragmentDirections.actionOrderListFragmentToConnectivityToolFragment()
         )
     }
